fix: await exit of FFmpeg process in `stopRecording` before returning
If FFmpeg is not awaited, Node might exit before it does; if, say,
everything is being run inside a Docker container, that could cause
FFmpeg to be killed before it can finish running, leaving an incomplete
video file and a whole lot of frames on disk. Now we wait for the
process to finish before returning from `stopRecording`, avoiding that
situation.

This also switches to synchronous writing of the image frames to avoid a
race condition with those completing and FFmpeg looking for them.